Renamed byteTransfer pacakage
