+++ runtime/executor/src/main/java/org/apache/nemo/runtime/executor/datatransfer/PipeInputReader.java
@@ -28,8 +28,8 @@ import org.apache.nemo.common.ir.vertex.IRVertex;
 import org.apache.nemo.runtime.common.RuntimeIdManager;
 import org.apache.nemo.runtime.common.plan.RuntimeEdge;
 import org.apache.nemo.runtime.executor.MetricMessageSender;
-import org.apache.nemo.runtime.executor.bytetransfer.ByteTransferContext;
-import org.apache.nemo.runtime.executor.bytetransfer.LocalOutputContext;
+import org.apache.nemo.runtime.executor.transfer.ByteTransferContext;
+import org.apache.nemo.runtime.executor.transfer.LocalOutputContext;
 import org.apache.nemo.runtime.executor.data.DataUtil;
 import org.apache.nemo.runtime.executor.data.PipeManagerWorker;
 import org.slf4j.Logger;
