+++ runtime/executor/src/main/java/org/apache/nemo/runtime/executor/datatransfer/PipeOutputWriter.java
@@ -24,8 +24,8 @@ import org.apache.nemo.common.punctuation.Watermark;
 import org.apache.nemo.runtime.common.RuntimeIdManager;
 import org.apache.nemo.runtime.common.plan.RuntimeEdge;
 import org.apache.nemo.runtime.common.plan.StageEdge;
-import org.apache.nemo.runtime.executor.bytetransfer.OutputContext;
-import org.apache.nemo.runtime.executor.bytetransfer.TransferOutputStream;
+import org.apache.nemo.runtime.executor.transfer.OutputContext;
+import org.apache.nemo.runtime.executor.transfer.TransferOutputStream;
 import org.apache.nemo.runtime.executor.data.PipeManagerWorker;
 import org.apache.nemo.runtime.executor.data.streamchainer.Serializer;
 import org.slf4j.Logger;
