+++ runtime/executor/src/main/java/org/apache/nemo/runtime/executor/transfer/ByteInputContext.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.nemo.runtime.executor.bytetransfer;
+package org.apache.nemo.runtime.executor.transfer;
 
 import io.netty.buffer.ByteBuf;
 import org.slf4j.Logger;
