+++ runtime/executor/src/main/java/org/apache/nemo/runtime/executor/transfer/package-info.java
@@ -19,4 +19,4 @@
 /**
  * Responsible for transferring block content from one executor to another.
  */
-package org.apache.nemo.runtime.executor.bytetransfer;
+package org.apache.nemo.runtime.executor.transfer;
